+++ CHANGES.md
@@ -26,6 +26,7 @@ Changes:
 - Retrieve lower court information in `wyo` #1569
 - Retrieve lower court information in `dc` #1569
 - Retrieve lower court information in `cal` scrapers #1569
+- Retrieve lower court information in 'alaska' #1569
 - Retrieve lower court information in `delaware` scrapers #1569
 
 Fixes:
@@ -42,6 +43,7 @@ Features:
 
 Changes:
 - Retrieve lower court information in `bap9` and `bap10` #1567
+- Retrieve lower court information in `ariz`  #1569
 
 Fixes:
 - fix `cafc` oral arguments scraper #1570